+++ ext/NeutralsUnitfulExt.jl
@@ -6,7 +6,7 @@ else
     using ..Neutrals, ..Unitful
 end
 
-using .Neutrals: is_dimensionless, is_static_number
+using .Neutrals: is_static_number
 using .Unitful: AbstractQuantity, Quantity, NoDims, unit, ustrip
 
 # Preserve units in multiplication and division.
@@ -24,9 +24,6 @@ Neutrals.impl_div(::Val{1}, x::Neutral, y::AbstractArray{<:AbstractQuantity}) =
 #
 #Neutrals._impl_div(x::Neutral{0}, y::AbstractArray{<:AbstractQuantity}) =
 #    similar(y, typeof(x/unit(eltype(y))))
-
-Neutrals.is_dimensionless(::Type{<:AbstractQuantity}) = false
-Neutrals.is_dimensionless(::Type{<:AbstractQuantity{<:Any,NoDims}}) = true
 
 Neutrals.is_static_number(::Type{<:AbstractQuantity{T}}) where {T} = is_static_number(T)