+++ NEWS.md
@@ -6,6 +6,11 @@ Versioning](https://semver.org).
 
 ## Unreleased
 
+### Changed
+
+- Non-exported `Neutrals.is_dimensionless` has been deprecated and replaced by
+  `TypeUtils.is_unitless`.
+
 ### Added
 
 - Non-exported public function `Neutrals.dispatch` and type `Neutrals.Dispatch` to mark